Ajout et modification de la fonction sound.c

+++ sound.h
@@ -2,7 +2,7 @@
 #define SOUND_H
 
 
-void getFrequency(int note, int* frequency, int* octave);
+void getFrequency(int note, float* frequency, int* octave);
 void* playNoteSound(void* arguments);
 
 #endif
